+++ src/views/components/Bookings/PaymentGateway/PaymentForm.tsx
@@ -5,7 +5,7 @@ import useAuth from "../../../../Hooks/useAuth/useAuth";
 import Button from "@mui/material/Button";
 import { Box } from "@mui/system";
 
-const PaymentForm = () => {
+const PaymentForm = ({ setDisplayLoader, setDisplay, bookingId,totalPrice }: any) => {
   const stripe = useStripe();
   const elements = useElements();
   const [error, setError] = useState(null);
@@ -13,14 +13,14 @@ const PaymentForm = () => {
   const cardElementOptions = {
     style: {
       base: {
-        fontSize: '16px',
-        color: '#424770',
-        '::placeholder': {
-          color: '#aab7c4',
+        fontSize: "16px",
+        color: "#424770",
+        "::placeholder": {
+          color: "#aab7c4",
         },
       },
       invalid: {
-        color: '#9e2146',
+        color: "#9e2146",
       },
     },
   };
@@ -39,19 +39,28 @@ const PaymentForm = () => {
       console.error(error);
     } else {
       // Send the token to your server
-
+      setDisplay(false);
+      setDisplayLoader(true);
       const response = await request.post("/order", {
         headers: {
           "Content-Type": "application/json",
         },
-        body: JSON.stringify({ token }),
+        body: {token:JSON.stringify({token}),price:totalPrice},
       });
 
       const result = await response.data;
 
       if (result.success) {
-        console.log("Payment successful:", result.charge);
+        const data = await request.post("/paymentSuccess", {
+          transactionId: result.charge.balance_transaction,
+          amount_captured: result.charge.amount_captured,
+          bookingId: bookingId,
+        });
+        setTimeout(() => {
+          setDisplayLoader(false);
+        });
       } else {
+        setDisplayLoader(false);
         console.error("Payment failed:", result.error);
       }
     }
@@ -60,8 +69,8 @@ const PaymentForm = () => {
   return (
     <form onSubmit={handleSubmit}>
       <label style={{ fontSize: "25px", margin: 10 }}>Enter Card details</label>
-      <Box sx={{border:'1px dashed lightgray',width:300,height:100,padding:5,m:2    }}   >
-        <CardElement  options={cardElementOptions} />
+      <Box sx={{ width: 350, padding: 5, m: 2 }}>
+        <CardElement options={cardElementOptions} />
       </Box>
       {error && <div style={{ color: "red" }}>{error}</div>}
       <Button variant="contained" type="submit" disabled={!stripe}>
