+++ src/views/components/Bookings/PaymentGateway/PaymentPage.tsx
@@ -1,15 +1,18 @@
 // App.js
-import React from 'react';
-import StripeProvider from './StripeProvider';
-import PaymentForm from './PaymentForm';
-import { Box } from '@mui/system';
+import React from "react";
+import StripeProvider from "./StripeProvider";
+import PaymentForm from "./PaymentForm";
+import { Box } from "@mui/system";
+import { Card, CardContent } from "@mui/material";
 
-const Booking = () => {
+const Booking = ({setDisplayLoader,setDisplay,bookingId,totalPrice}:any) => {
   return (
     <StripeProvider>
-      <Box sx={{width:400,border:'1px solid lightgray',padding:2}} >
-        <PaymentForm />
-      </Box>
+     
+          <Box sx={{ width: 400,padding: 2,border:'1px solid lightgray' }}  >
+            <PaymentForm setDisplayLoader={setDisplayLoader} setDisplay={setDisplay} bookingId={bookingId} totalPrice={totalPrice} />
+          </Box>
+    
     </StripeProvider>
   );
 };
